src/requirements.txt
dbus-python==1.2.18
pyperclip==1.8.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
vext==0.7.6
vext.gi==0.7.4
